ftrace: use quint32 for context counter
Some old platforms don't have atomic integer supports for 64-bit wise
integer.

As indicated in GPUVis:

	// Erase all knowledge of this ctx so it can be reused

We can reuse numbers after end_ctx, so with we should be fine with
32-bit integer.

Let's use a 32 bit integer for context.

+++ src/ftrace.h
@@ -73,7 +73,7 @@ class KWIN_EXPORT FTraceDuration
 public:
     template<typename... Args> FTraceDuration(Args... args)
     {
-        static QAtomicInteger<qulonglong> s_context = 0;
+        static QAtomicInteger<quint32> s_context = 0;
         QTextStream stream(&m_message);
         (stream << ... << args);
         stream.flush();
@@ -85,7 +85,7 @@ public:
 
 private:
     QByteArray m_message;
-    qulonglong m_context;
+    quint32 m_context;
 };
 
 } // namespace KWin
